0x14-bit_manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0, Xor;
+	unsigned long int count = 0, Xor;
 
 	Xor = n ^ m;
 	for (; Xor; Xor >>= 1)
